refactored userpartylist into 4 components
