+++ client/components/ItemList.js
@@ -13,7 +13,7 @@ import PropTypes from 'prop-types'
 import AddCircleOutlineIcon from '@material-ui/icons/AddCircleOutline'
 import RemoveCircleOutlineIcon from '@material-ui/icons/RemoveCircleOutline'
 import Card from '@material-ui/core/Card'
-import {AddItem} from '../components'
+import {AddItem} from '.'
 import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails'
 
 import {putItem} from '../store'
