+++ client/components/UpcomingParty.js
@@ -1,11 +1,9 @@
-import React, {Component} from 'react'
-import {connect} from 'react-redux'
+import React from 'react'
 import moment from 'moment'
 import {Link} from 'react-router-dom'
 
 import {withStyles} from '@material-ui/core/styles'
 import Card from '@material-ui/core/Card'
-import CardHeader from '@material-ui/core/CardHeader'
 import CardMedia from '@material-ui/core/CardMedia'
 import CardContent from '@material-ui/core/CardContent'
 import ExpansionPanel from '@material-ui/core/ExpansionPanel'
@@ -13,10 +11,14 @@ import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary'
 import Grid from '@material-ui/core/Grid'
 import Typography from '@material-ui/core/Typography'
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
+import Button from '@material-ui/core/Button'
 
 const styles = theme => ({
   heading: {
     fontSize: theme.typography.pxToRem(15)
+  },
+  column: {
+    flexBasis: '50'
   },
   media: {
     height: 0,
@@ -24,7 +26,7 @@ const styles = theme => ({
   }
 })
 
-const UpcomingEvent = props => {
+const UpcomingParty = props => {
   const {upcomingEvent, user, classes} = props
 
   return upcomingEvent.id ? (
@@ -58,12 +60,6 @@ const UpcomingEvent = props => {
         to={`/parties/${upcomingEvent.id}`}
         style={{textDecoration: 'none'}}
       >
-        {/* <CardHeader
-          title={parties.upcomingEvent.title}
-          subheader={moment(parties.upcomingEvent.date).format(
-            'LLLL'
-          )}
-        /> */}
         <CardMedia className={classes.media} image={upcomingEvent.imageUrl} />
         <CardContent>
           <Typography component="p">{upcomingEvent.description}</Typography>
@@ -72,10 +68,24 @@ const UpcomingEvent = props => {
     </ExpansionPanel>
   ) : (
     <ExpansionPanel defaultExpanded>
-      <ExpansionPanelSummary expandIcon={<ExpandMoreIcon />}>
-        <Typography color="inherit">
-          No upcoming feasts... why not create one?
-        </Typography>
+      <ExpansionPanelSummary>
+        <Grid container justify="space-between">
+          <Grid item>
+            <Typography color="inherit">No upcoming feasts.</Typography>
+            <Typography color="inherit">Why not create one?</Typography>
+          </Grid>
+          <Grid item>
+            <Button
+              component={Link}
+              to="/addparty"
+              className={classes.button}
+              variant="contained"
+              color="primary"
+            >
+              Feast!
+            </Button>
+          </Grid>
+        </Grid>
       </ExpansionPanelSummary>
       <Card className={classes.card}>
         <CardMedia
@@ -87,4 +97,4 @@ const UpcomingEvent = props => {
   )
 }
 
-export default withStyles(styles)(UpcomingEvent)
+export default withStyles(styles)(UpcomingParty)
