+++ client/components/index.js
@@ -14,4 +14,7 @@ export {default as UserPartyList} from './UserPartyList'
 export {default as AddParty} from './AddParty'
 export {default as EditParty} from './EditParty'
 export {default as Gallery} from './Gallery'
-export {default as UpcomingEvent} from './UpcomingEvent'
+export {default as UpcomingParty} from './UpcomingParty'
+export {default as HostingParties} from './HostingParties'
+export {default as AttendingParties} from './AttendingParties'
+export {default as PastParties} from './PastParties'
